morello: Add experimental version of the revoker with direct DIRTY-CLEAN transitions.
Transitioning to clean directly is complicated. This patch restricts transitions
only for pages that do not have aliases and have been xbusied.
The transition involves two stages:
1. The PTE of a presumed cap-clean page is moved to trap on capability store.
2. Scan again the page
3. If the page is still cap-clean and the PTE is still in the clean state,
then fully IDLE the page by stripping PGA_CAPSTORE.

This seem to have a fundamental problem because we can't move the PTE to
any of the CLEAN / DIRTIABLE states unless we know for sure that there
are no capabilities. Otherwise a concurrent thread may write a non-revoked
capability, race with the revoker and load it again but find it untagged.