morello: Implement batch-clean revoker.
This revoker does not transition cap-clean pages to the CLEAN
state concurrently. Instead, the transition is deferred to the next
stop-the-world phase, during which concurrent writes are not possible.
In this way, the secondary page scan can determine whether the page
is really cap-clean and commit to the CLEAN state transition.
This is only done for non-aliasing pages, otherwise a page may end up
on multiple revoker clean lists, which is not supported.
There are further obstacles to having this for aliasing pages, as we
need to synchronise the PTE bits for all mappings when moving the
page to the CLEAN state.