Simplify registration of the "native" vDSO timekeep
Allow the first non-COMPAT32 sysentvec with SV_TIMEKEEP set to allocate
the "native" timekeep.  The previous code was too fragile as it would
panic if multiple ABIs tried to register on because the "main" ABI was
registered.  This happens on CheriBSD when we have several sysentvec and
it could happen if linkage order caused Linux to be ordered first.

Effort:		CHERI upstreaming
Sponsored by:	Innovate UK, DARPA